+++ src/main/java/org/folio/linked/data/integration/kafka/sender/search/WorkCreateMessageSender.java
@@ -7,8 +7,8 @@ import static java.util.Collections.singletonList;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.INSTANCE;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.WORK;
 import static org.folio.linked.data.domain.dto.ResourceIndexEventType.CREATE;
-import static org.folio.linked.data.util.BibframeUtils.extractWorkFromInstance;
 import static org.folio.linked.data.util.Constants.FOLIO_PROFILE;
+import static org.folio.linked.data.util.ResourceUtils.extractWorkFromInstance;
 
 import java.util.Collection;
 import java.util.List;
