+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/instance/sub/SupplementaryContentMapperUnit.java
@@ -1,8 +1,8 @@
 package org.folio.linked.data.mapper.dto.monograph.instance.sub;
 
 import static org.folio.ld.dictionary.ResourceTypeDictionary.SUPPLEMENTARY_CONTENT;
-import static org.folio.linked.data.util.BibframeUtils.getFirstValue;
-import static org.folio.linked.data.util.BibframeUtils.putProperty;
+import static org.folio.linked.data.util.ResourceUtils.getFirstValue;
+import static org.folio.linked.data.util.ResourceUtils.putProperty;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import java.util.HashMap;
