+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/work/sub/ReferenceMapperUnit.java
@@ -1,7 +1,7 @@
 package org.folio.linked.data.mapper.dto.monograph.work.sub;
 
-import static org.folio.linked.data.util.BibframeUtils.ensureLatestReplaced;
-import static org.folio.linked.data.util.BibframeUtils.fetchResource;
+import static org.folio.linked.data.util.ResourceUtils.ensureLatestReplaced;
+import static org.folio.linked.data.util.ResourceUtils.fetchResource;
 
 import java.util.function.BiConsumer;
 import lombok.RequiredArgsConstructor;
